27/05/24

+++ src/components/product/productTable/Table.jsx
@@ -1,162 +1,3 @@
-// import React, { useState, useEffect } from "react";
-// import axios from "axios";
-// import Modal from "../../modals/Modal";
-// import AddProduct from "../AddProduct";
-
-// function Table() {
-//   const [products, setProducts] = useState([]);
-//   const [singleProduct, setsingleProduct] = useState(null);
-//   const [refresh, setRefresh] = useState(true);
-//   const [isModalOpen, setIsModalOpen] = useState(false);
-
-//   const fetchProducts = async () => {
-//     try {
-//       const { data } = await axios.get("http://localhost:3000/products/getall");
-//       if (data.success) {
-//         setProducts(data.products);
-//       }
-//     } catch (error) {
-//       console.error(error);
-//     }
-//   };
-
-//   useEffect(() => {
-//     fetchProducts();
-    
-//   }, [refresh]);
-
-//   // useEffect(() => {
-
-//   //   handleCloseModal
-//   // }, [products])
-  
-
-//   async function handleDelete(id) {
-//     try {
-//       const { data } = await axios.delete(
-//         `http://localhost:3000/products/delete/${id}`
-//       );
-//       if (data.success) {
-//         console.log("deleted");
-//         setRefresh((prev) => !prev);
-//       }
-//     } catch (error) {
-//       console.log(error);
-//       console.log("not deleted");
-//     }
-//   }
-
-//   const handleEdit = (data) => {
-    
-//     setsingleProduct(data);
-//     setIsModalOpen(true);
-//   };
-
-//   const handleCloseModal = () => {
-//     setsingleProduct(null);
-//     setIsModalOpen(false);
-//   };
-
-//   const handleSaveProduct = () => {
-//     handleCloseModal();
-//   };
-  
-
-//   return (
-//     <div>
-//       <button
-//         type="btn"
-//         className="justify-self-end text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800:"
-//         onClick={() => handleEdit(null)}
-//       >
-//         Add Product
-//       </button>
-
-//       <div className="relative w-[80%] mx-auto overflow-x-auto shadow-md sm:rounded-lg">
-//         <table className="w-full text-sm h-full text-left rtl:text-right text-gray-500 dark:text-gray-400">
-//           <thead className="text-xs text-gray-700 uppercase bg-gray-50 dark:bg-gray-700 dark:text-gray-400">
-//             <tr>
-//               <th scope="col" className="px-6 py-3">
-//                 Company
-//               </th>
-//               <th scope="col" className="px-6 py-3">
-//                 Model
-//               </th>
-//               <th scope="col" className="px-6 py-3">
-//                 Colors
-//               </th>
-//               <th scope="col" className="px-6 py-3">
-//                 Car Type
-//               </th>
-//               <th scope="col" className="px-6 py-3">
-//                 Year
-//               </th>
-//               <th scope="col" className="px-6 py-3">
-//                 Price
-//               </th>
-//               <th scope="col" className="px-6 py-3">
-//                 Image
-//               </th>
-//               <th scope="col" className="px-6 py-3">
-//                 Actions
-//               </th>
-//             </tr>
-//           </thead>
-//           <tbody>
-//             {products.map((product) => (
-//               <tr
-//                 key={product._id}
-//                 className="bg-white border-b dark:bg-gray-900 dark:border-gray-700"
-//               >
-//                 <td className="px-6 py-4">{product.company}</td>
-//                 <td className="px-6 py-4">{product.model}</td>
-//                 <td className="px-6 py-4">{product.colors.join(", ")}</td>
-//                 <td className="px-6 py-4">{product.car_type}</td>
-//                 <td className="px-6 py-4">{product.year}</td>
-//                 <td className="px-6 py-4">{product.price}</td>
-//                 <td className="px-6 py-4">
-//                   <img
-//                     width={60}
-//                     height={60}
-//                     src={product.image_link}
-//                     alt={product.model}
-//                   />
-//                 </td>
-//                 <td className="px-6 py-4 flex space-x-2">
-//                   <button
-//                     className="bg-blue-500 text-white px-3 py-1 rounded-md hover:bg-blue-600"
-//                     onClick={() => handleEdit(product)}
-//                   >
-//                     Edit
-//                   </button>
-//                   <button
-//                     className="bg-red-500 text-white px-3 py-1 rounded-md hover:bg-red-600"
-//                     onClick={() => handleDelete(product._id)}
-//                   >
-//                     Remove
-//                   </button>
-//                 </td>
-//               </tr>
-//             ))}
-//           </tbody>
-//         </table>
-//       </div>
-//       {/* <footer>
-//         <button onClick={() => console.log(singleProduct)} className="bg-green-500 text-white px-3 py-1 rounded-md hover:bg-green-600 mt-4">
-//           Print Products
-//         </button>
-//       </footer> */}
-//       <Modal isOpen={isModalOpen} onClose={handleCloseModal}>
-//         <AddProduct product={singleProduct} setRefresh={setRefresh} onSave={handleSaveProduct}/>
-//       </Modal>
-//     </div>
-//   );
-// }
-// // {/* onSave={handleSaveProduct} */}
-
-// export default Table;
-
-// ****************************************
 
 import React, { useState, useEffect } from "react";
 import axios from "axios";
@@ -164,13 +5,16 @@ import Modal from "../../modals/Modal";
 import AddProduct from "../AddProduct";
 import { AiOutlineEye } from "react-icons/ai";
 import ProductPage from "./ProductPage";
+import { FiRefreshCcw } from "react-icons/fi";
+
 
 function Table() {
   const [products, setProducts] = useState([]);
   const [singleProduct, setsingleProduct] = useState(null);
   const [refresh, setRefresh] = useState(true);
-  const [isModalOpen, setIsModalOpen] = useState(false);
+  const [isProductModalOpen, setisProductModalOpen] = useState(false);
   const [isProductPageOpen, setIsProductPageOpen] = useState(false);
+
 
   const fetchProducts = async () => {
     try {
@@ -189,27 +33,30 @@ function Table() {
 
   async function handleDelete(id) {
     try {
-      const { data } = await axios.delete(
-        `http://localhost:3000/products/delete/${id}`
-      );
-      if (data.success) {
-        console.log("deleted");
-        setRefresh((prev) => !prev);
-      }
+        const sure = confirm("Are you sure you want to delete?");
+        if (sure) {
+            const { data } = await axios.delete(
+                `http://localhost:3000/products/delete/${id}`
+            );
+            if (data.success) {
+                console.log("Product deleted successfully");
+                setRefresh((prev) => !prev);
+            }
+        }
     } catch (error) {
-      console.log(error);
-      console.log("not deleted");
+        console.log("Error deleting product:", error);
     }
-  }
+}
+
 
   const handleEdit = (data) => {
     setsingleProduct(data);
-    setIsModalOpen(true);
+    setisProductModalOpen(true);
   };
 
   const handleCloseModal = () => {
     setsingleProduct(null);
-    setIsModalOpen(false);
+    setisProductModalOpen(false);
   };
 
   const handleSaveProduct = () => {
@@ -226,6 +73,8 @@ function Table() {
     setIsProductPageOpen(false);
   };
 
+  
+
   return (
     <div>
       <button
@@ -234,6 +83,13 @@ function Table() {
         onClick={() => handleEdit(null)}
       >
         Add Product
+      </button>
+      <button
+        type="btn"
+        className="justify-self-end text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800"
+        onClick={() => setRefresh((prev) => !prev)}
+      >
+        <FiRefreshCcw />
       </button>
 
       <div className="relative w-[80%] mx-auto overflow-x-auto shadow-md sm:rounded-lg">
@@ -303,6 +159,7 @@ function Table() {
                     className="bg-green-500 text-white px-3 py-1 rounded-md hover:bg-green-600"
                     onClick={() => handleViewProduct(product)}
                   >
+                    {/* כפתור צפיית מוצר */}
                     <AiOutlineEye />
                   </button>
                 </td>
@@ -312,7 +169,7 @@ function Table() {
         </table>
       </div>
 
-      <Modal isOpen={isModalOpen} onClose={handleCloseModal}>
+      <Modal isOpen={isProductModalOpen} onClose={handleCloseModal}>
         <AddProduct product={singleProduct} setRefresh={setRefresh} onSave={handleSaveProduct} />
       </Modal>
 
@@ -321,6 +178,7 @@ function Table() {
           <ProductPage product={singleProduct} />
         </Modal>
       )}
+      
     </div>
   );
 }
